+++ bindings/nodejs/types/wallet/bridge/account.ts
@@ -267,14 +267,6 @@ export type __RegisterParticipationEventsMethod__ = {
     };
 };
 
-export type __RequestFundsFromFaucetMethod__ = {
-    name: 'requestFundsFromFaucet';
-    data: {
-        url: string;
-        address: string;
-    };
-};
-
 export type __RetryTransactionUntilIncludedMethod__ = {
     name: 'retryTransactionUntilIncluded';
     data: {