+++ sdk/tests/wallet/accounts.rs
@@ -153,11 +153,13 @@ async fn account_rename_alias() -> Result<()> {
     let account = wallet.create_account().with_alias("Alice".to_string()).finish().await?;
 
     assert_eq!(account.alias().await, "Alice".to_string());
+    assert_eq!(account.read().await.alias(), "Alice");
 
     // rename account
     account.set_alias("Bob").await?;
 
     assert_eq!(account.alias().await, "Bob".to_string());
+    assert_eq!(account.read().await.alias(), "Bob");
 
     tear_down(storage_path)
 }